requirements.txt
notion-client
python-dotenv
clipboard
